+++ Projects/Metadata_Extractor/metadata.pyw
@@ -40,6 +40,9 @@ output_path = None
 #For debugging only, manually set
 debug = False
 #Gets called on set directory button press to set image directory
+def reset():
+    used.clear()
+    results.clear()
 def set_dir():
     global directory_path
     directory_path = Path(filedialog.askdirectory(title="Select a Directory"))
@@ -56,7 +59,7 @@ def scan_image(directory_path, files_and_dirs, i, list_of_dates, list_of_images)
         raise ValueError(f"No metadata found for {filename}")
     list_of_dates.append(dt) #Add to lists
     list_of_images.append(filename)
-def find_directories(directory_path):
+def find_directories(directory_path, final_output):
     global current_sheet_name
     name_index = 0
     directories_found = os.listdir(directory_path) #Only one level of recursion
@@ -74,28 +77,29 @@ def find_directories(directory_path):
     for directory_path in directories_to_scan:
         current_sheet_name = sheet_names[name_index]
         name_index+=1
-        sort_and_extract_irregular(directory_path, True)
-    create_sheet(output_path, True, None)
+        sort_and_extract_irregular(directory_path, True, final_output)
+    create_sheet(final_output, True, None)
 def check_recursive(directory_path):
     global output_path
     if debug: #No need to set output_path every time
-                output_path = Path.home() / "Downloads" / "output.xlsx"
+                output_path = Path.home() / "Downloads"
     if directory_path == None or output_path == None:
             messagebox.showerror( #If no output/directory path was set, error to allow for changes
                         "Invalid path(s) specified.", message="Please set an image/video directory path OR output path.")
             return
+    reset()
     if name_entry.get() != "Filename (default output.xlsx)":
-        output_path = Path(output_path) / clean_name(name_entry.get()) # type: ignore the output_path is None error
+        final_output = Path(output_path) / clean_name(name_entry.get())
     else:
-        output_path = Path(output_path) / "output.xlsx" # type: ignore the output_path is None error
+        final_output = Path(output_path) / "output.xlsx"
     recursive = recursive_var.get()
     if recursive:
-        find_directories(directory_path)
+        find_directories(directory_path, final_output)
     else:
         with os.scandir(directory_path) as entries:
                 photo_count = sum(1 for entry in entries if entry.is_file() and entry.name.lower().endswith((photo_types)))
                 if photo_count > 0:
-                    sort_and_extract_irregular(directory_path, False)
+                    sort_and_extract_irregular(directory_path, False, final_output)
                 else:
                         messagebox.showerror("Empty directory", message="No files were found. Please select a valid directory")
                         return
@@ -134,7 +138,7 @@ def clean_name(name):
         name += ".xlsx"
     return name
 #Gets called initially, assumes irregular pattern, calls sort_and_extract() if regular
-def sort_and_extract_irregular(directory_path, recursive):
+def sort_and_extract_irregular(directory_path, recursive, final_output):
     global output_series
     output_series = True
     list_of_dates = []
@@ -174,7 +178,7 @@ def sort_and_extract_irregular(directory_path, recursive):
                             photos_before_video_updated = True
                             photos_before_video = 0
                 if (i == 0 and photos_before_video == 0) or output_series == False:
-                    sort_and_extract(directory_path,recursive)
+                    sort_and_extract(directory_path,recursive, final_output)
                     return
                 if photos_before_video_updated and regular_pattern_found == False:
                     end = False
@@ -252,11 +256,11 @@ def sort_and_extract_irregular(directory_path, recursive):
     fixed_date, fixed_time = fix_time(list_of_dates)
     df = create_dataframe(img_series, list_of_images, fixed_date, fixed_time, recursive, directory_path)
     if not recursive:
-        create_sheet(output_path, recursive, df)
+        create_sheet(final_output, recursive, df)
 #Gets called on "Execute" button press 
 #Paths supplied by user or debug mode
 #Assumes regular pattern
-def sort_and_extract(directory_path, recursive):
+def sort_and_extract(directory_path, recursive, final_output):
     global output_series
     list_of_dates = []
     list_of_images = []
@@ -314,7 +318,7 @@ def sort_and_extract(directory_path, recursive):
     img_series, final_images, final_dates, final_times = make_series(list_of_images,photos_before_video_updated,photos_before_video,fixed_date,fixed_time)
     df = create_dataframe(img_series, final_images, final_dates, final_times, recursive, directory_path)
     if not recursive:
-        create_sheet(output_path, recursive, df)
+        create_sheet(final_output, recursive, df)
 
 #For regular patterns only
 def make_series(list_of_images, photos_before_video_updated, photos_before_video, fixed_date, fixed_time):
@@ -422,31 +426,51 @@ def create_sheet(output_path, recursive, df):
             out_file = Path(output_path) #Output 
             if not results:
                 messagebox.showerror("No data", message="No folders with photos were found.")
-                root.destroy()
-                sys.exit()
+                if not keepopen_var.get():
+                    root.destroy()
+                    sys.exit()
+                return
             if recursive:
-                with pd.ExcelWriter(output_path) as writer:
-                    for df, sheet_name in results:
-                        df.to_excel(writer, sheet_name=sheet_name, index=False)
+                if output_path.exists():
+                    overwrite_check = messagebox.askyesno("Overwrite?", message=output_path.as_posix() + " \nalready exists. Overwrite?")
+                    if overwrite_check:
+                        with pd.ExcelWriter(output_path) as writer:
+                            for df, sheet_name in results:
+                                df.to_excel(writer, sheet_name=sheet_name, index=False)
+                        messagebox.showinfo("Success!", message="Success, outputted at " + output_path.as_posix()) #Tell user where file outputted
+                else:
+                    with pd.ExcelWriter(output_path) as writer:
+                        for df, sheet_name in results:
+                            df.to_excel(writer, sheet_name=sheet_name, index=False)
+                    messagebox.showinfo("Success!", message="Success, outputted at " + output_path.as_posix()) #Tell user where file outputted
             else:
-                df.to_excel(out_file, sheet_name="Output", index=False) #Make an Excel file for importing
-            messagebox.showinfo(
-                "Success!", message="Success, outputted at " + output_path.as_posix() #Tell user where file outputted
-            )
-            root.destroy() #End program
-            sys.exit()
+                if output_path.exists():
+                    overwrite_check = messagebox.askyesno("Overwrite?", message=output_path.as_posix() + " \nalready exists. Overwrite?")
+                    if overwrite_check:
+                        df.to_excel(out_file, sheet_name="Output", index=False) #Make an Excel file for importing
+                        messagebox.showinfo("Success!", message="Success, outputted at " + output_path.as_posix()) #Tell user where file outputted
+                else:
+                    df.to_excel(out_file, sheet_name="Output", index=False) #Make an Excel file for importing
+                    messagebox.showinfo("Success!", message="Success, outputted at " + output_path.as_posix()) #Tell user where file outputted
+            if not keepopen_var.get():
+                if output_path.exists() and keepopen_var.get() == False:
+                    messagebox.showinfo("File not overwritten.", message="File not overwritten. Cancelled successfully.") #Tell user where file outputted
+                root.destroy() #End program
+                sys.exit()
         except Exception as e:
             messagebox.showerror( #If no output path was set, error and close
-                "An error occurred.", message="Is another program is using output.xlsx? Error caught: " + str(e)
-            )
-            root.destroy() #End program
-            sys.exit()
+                "An error occurred.", message="Is another program is using output.xlsx? Error caught: " + str(e))
+            if not keepopen_var.get():
+                root.destroy() #End program
+                sys.exit()
 
 root = tk.Tk() #Tkinter root
 frame = tk.Frame(root) #Tkinter frame window
 execute_button = tk.Button(frame, command=lambda:check_recursive(directory_path), text="Execute") #Buttons to click
 set_directory_button = tk.Button(frame, command=set_dir, text="Set folder with images")
 set_output_button = tk.Button(frame, command=set_output, text="Set output folder")
+keepopen_var = tk.BooleanVar(value=True)  # default checked
+keepopen_checkbox = tk.Checkbutton(frame, text="Keep open?", variable=keepopen_var)
 recursive_var = tk.BooleanVar(value=False)  # default unchecked
 recursive_checkbox = tk.Checkbutton(frame, text="Scan subfolders", variable=recursive_var)
 name_entry = tk.Entry(frame, width=30, foreground="gray")
@@ -459,5 +483,6 @@ set_output_button.pack()
 name_entry.pack()
 recursive_checkbox.pack()
 execute_button.pack()
+keepopen_checkbox.pack()
 root.geometry("200x120") #Set dimensions of window to open
 root.mainloop() #Ensure window only closes by user choice
